Big Update
I have added a main menu lobby scene which allows the players to chose host, join or exit the game. During the game, the players are able to win by getting a set amount of kills and then they return to the lobby after a few seconds. Also added some post processing and animations for the first person view of the gun